投稿したユーザー以外はdestroyできない

+++ app/controllers/items_controller.rb
@@ -35,6 +35,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to action: :index unless @item.user_id == current_user.id
     redirect_to action: :index if @item.destroy
   end
